refactor(send-fee): viewModel 분리
- isMaxMode 변수를 ViewModel로 이전하였습니다.
- 이미 viewModel에서 관리하고 있는 _confirmedBalance 변수는 view에서 삭제했습니다.
- FeeSelectionItem->FeeSelectionItemCard, widgets폴더 하위 파일로 이동하였습니다.
#68
